# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "For the administration of analytics operators.",
        "title": "Analytics-Operator-Repo-V2 API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "{version}"
    },
    "basePath": "/",
    "paths": {
        "/operator": {
            "get": {
                "description": "Gets all operators",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get operators",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Maximum number of operators to return, at most 1000; 0 means no limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of operators to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name:desc",
                        "description": "Sort as field[:asc|desc]; only name is sortable",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Case-sensitive substring match on name; treated literally",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.OperatorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes multiple operators",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Delete multiple operators",
                "parameters": [
                    {
                        "description": "ID list",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/": {
            "put": {
                "description": "Stores an operator",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Create operator",
                "parameters": [
                    {
                        "description": "Create operator",
                        "name": "operator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Operator"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operator/{id}": {
            "get": {
                "description": "Gets a single operator",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Operator"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Validates and updates an operator",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Update operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update operator",
                        "name": "operator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Operator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an operator",
                "tags": [
                    "Operator"
                ],
                "summary": "Delete operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lib.Operator": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "config_values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Value"
                    }
                },
                "cost": {
                    "type": "integer"
                },
                "dateCreated": {
                    "type": "string"
                },
                "dateUpdated": {
                    "type": "string"
                },
                "deploymentType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Value"
                    }
                },
                "name": {
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Value"
                    }
                },
                "pub": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "lib.OperatorResponse": {
            "type": "object",
            "properties": {
                "operators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Operator"
                    }
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "lib.Value": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}